Sync to EN 44169:
I think the reference here is a little mistake. We talk
about \ on the line above the changed line because
Windows uses that.  We then state that FreeBSD does not
use that, but with our Unix Slash (/) in the example.

My feeling is that this should have been the Windows
slash to make it clear that this does not work and
should not be used on FreeBSD (or most other Unixes
for that matter).

Facilitated by:     Snow B.V.
Found:              During translations.